feat(404): restructure routes
