+++ src/routes/$lang/error/error.styles.scss
@@ -1,0 +1,58 @@
+@import 'src/styles/utils/_utils.scss';
+
+.catch-all {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  background: url("data:image/svg+xml,%3Csvg width='784' height='1078' viewBox='0 0 784 1078' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cmask id='mask0_435_260' style='mask-type:alpha' maskUnits='userSpaceOnUse' x='0' y='0' width='862' height='1078'%3E%3Cellipse cx='430.963' cy='539.045' rx='538.955' ry='430.654' transform='rotate(-90 430.963 539.045)' fill='url(%23paint0_radial_435_260)' /%3E%3C/mask%3E%3Cg mask='url(%23mask0_435_260)'%3E%3Cpath d='M752.683 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M668.325 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M583.96 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M499.595 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M415.237 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M330.872 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M246.513 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M162.149 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M77.7897 -25.1733V1102.55' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 68.8045H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 162.779H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 256.758H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 350.731H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 444.71H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 538.683H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 632.662H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 726.637H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 820.615H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 914.594H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3Cpath d='M-175.298 1008.57H837.048' stroke='white' stroke-opacity='0.15' stroke-linecap='round' stroke-linejoin='round' /%3E%3C/g%3E%3Cdefs%3E%3CradialGradient id='paint0_radial_435_260' cx='0' cy='0' r='1' gradientUnits='userSpaceOnUse' gradientTransform='translate(430.963 539.045) rotate(90) scale(430.654 538.955)'%3E%3Cstop stop-color='%23C4C4C4' /%3E%3Cstop offset='1' stop-color='%23C4C4C4' stop-opacity='0' /%3E%3C/radialGradient%3E%3C/defs%3E%3C/svg%3E")
+      space repeat center,
+    theme('background');
+
+  &__main {
+    color: theme('textColor');
+    font-family: 'Open Sans';
+
+    > * {
+      text-align: center;
+    }
+
+    h1 {
+      font-size: 20rem;
+      background: theme('primaryGradient');
+      background-clip: text;
+      -webkit-background-clip: text;
+      -webkit-text-fill-color: transparent;
+      -webkit-text-stroke-width: 1px;
+      -webkit-text-stroke-color: theme('textColor');
+      line-height: 18rem;
+    }
+
+    h2 {
+      font-size: 5rem;
+    }
+
+    h3 {
+      font-size: 3rem;
+    }
+  }
+}
+
+// @media (max-width: $deviceWidth) {
+//   .catch-all__social-media {
+//     width: 100vw;
+//     height: 10%;
+//   }
+// }
+
+// @media (min-width: $deviceWidth) {
+//   .catch-all__social-media {
+//     position: absolute;
+//     bottom: 0;
+//     right: 0;
+//     width: clamp(60px, 3.5vw, 120px);
+//     height: clamp(200px, 25%, 550px);
+//   }
+// }
