--- src/routes/CatchAll/catchAll.styles.scss
@@ -1,24 +1,0 @@
-@import 'src/styles/utils/_utils.scss';
-
-.catch-all {
-  position: relative;
-  width: 100%;
-  height: 100%;
-}
-
-@media (max-width: $deviceWidth) {
-  .catch-all__social-media {
-    width: 100vw;
-    height: 10%;
-  }
-}
-
-@media (min-width: $deviceWidth) {
-  .catch-all__social-media {
-    position: absolute;
-    bottom: 0;
-    right: 0;
-    width: clamp(60px, 3.5vw, 120px);
-    height: clamp(200px, 25%, 550px);
-  }
-}
